Renamed "Web_TiledLayer" to remove the underscore. The name "WebTiledLayer" conflicts with the type with the same name, so renamed it "LoadWebTiledLayer".
