+++ src/Forms/Shared/Samples/Layers/LoadWebTiledLayer/LoadWebTiledLayer.xaml
@@ -3,7 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"             
              xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Xamarin.Forms;assembly=Esri.ArcGISRuntime.Xamarin.Forms"
              xmlns:mapping="clr-namespace:Esri.ArcGISRuntime.Mapping;assembly=Esri.ArcGISRuntime" 
-             x:Class="ArcGISRuntime.Samples.Web_TiledLayer.Web_TiledLayer">
+             x:Class="ArcGISRuntime.Samples.LoadWebTiledLayer.LoadWebTiledLayer">
   <Grid>
     <esriUI:MapView x:Name="MyMapView"/>
   </Grid>
