+++ src/UWP/ArcGISRuntime.UWP.Viewer/Samples/Layers/LoadWebTiledLayer/LoadWebTiledLayer.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl 
-    x:Class="ArcGISRuntime.UWP.Samples.Web_TiledLayer.Web_TiledLayer"
+    x:Class="ArcGISRuntime.UWP.Samples.LoadWebTiledLayer.LoadWebTiledLayer"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
