+++ src/WPF/ArcGISRuntime.WPF.Viewer/Samples/Layers/LoadWebTiledLayer/LoadWebTiledLayer.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGISRuntime.WPF.Samples.Web_TiledLayer.Web_TiledLayer"
+﻿<UserControl x:Class="ArcGISRuntime.WPF.Samples.LoadWebTiledLayer.LoadWebTiledLayer"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
